Initial lag v shift data processing script finished, bugs fixed
Fixed issues in the arrangement of the data in the versus nonlinearity
in the observations and the shift for better consistency with all other
processing scripts.  Added comments, need to revise plotting scripts for
the new formats.  Initial version of processing lag v shift data appears
to be working, needs further testing.