Fixed bug in SIETKS MDA, shift > 1
The weights scheme between all shift > 1 MDA are identical, but
I noticed an issue with the posterior statistics looked incorrect based
on diagnostics.  The MDA scheme was inadvertently re-analyzing filtered
states with the same transform twice, this has been fixed and the
re-analysis scheme in the balancing stage has been shortened in terms
of redundant steps for greater speed.