Running initial revision of the parallelism for Pronghorn
After a few attempts to run the built in Julia parallelism robustly to
failed sub-experiments in a batch submit, it seems like the Slurm
scheduling of the jobs is a more robust solution, especially in the way
that one can monitor the current jobs, their progress and complete /
incomplete runs.  Unfortunately this means that on the common queue this
fills up the queue list rapidly with job names, but they are all
associated to the same account which is separate, so this is just an
aesthetic inconvenience.