Restructuring repo
When packaging and importing the unregistered and dev versions I lost
the ability to import the SingleExperimentDriver.jl despite appending
the LOAD_PATH.  It appears to be fixed by moving this to the experiments
directory, which is propbably preferrable along with the other clean up
of slurm scripts.  Pushing initial commit to test on Caprine / Okapi /
Pronghorn.