Revising relative paths for experiments
Experiments should be run from the slurm_submit_script directory or the
experiment directory with the SingleExperimentDriver in order to get the
relative paths to the usual data directories for the outputs.  Want to
investigate ways to add relative paths from the root of a project --
need to see if this exists.