Cleaning up test cases
Starting to work on updates to the merged branch, want to update
this to run all test cases together in the runtests.jl module when
instantiating the test command.  Preparing to add additional test
cases in the main module, will merge this later with additional
work to the time series generation test.