Minor reductions of flops in RK scheme, continuing development of IEEE
Removing minor extra flop counts in the RK scheme that existed in the
deterministic formulation.  Also pre-allocating the space for the
coefficients for slightly improved speed.  We appear to match the
synchronous state in the Python simulations formerly in the asymptotic
solution with the Julia model now, though the noise perturbed version
seems to still have a qualtitatively different behavior.  This is going
through continuing testing.