Core functionality change
For parameter estimation problems and the need for flexible assignment
of the dx_parameters therein, this is being revised so that dx_params
will take the form of a dictionary where the derivative functions will
call the necessary parameters by name.  This is merged with the extended
state formalism such that the parameter sample dictionary will need to
be included in the kwargs; this is to specify the parameter name and the
indices of the extended state vector that will be merged into the
dx_params dict for the purposes of the ensemble integration.

All functionality appears to be working with the GenerateTimeSeries and
FilterExperiments, but this is pending integration into the smoother
experiments.  Various other changes will need to be propagated into,
e.g., test cases for the code where the types of inputs will now break
these scripts.