Major refactor of the code base
Revised the structure for complete module heirarchy through the core
code in src, also changed core utilities like JLD to JLD2 in order to
handle bugs.  Changed path names largely to ones relative to the file in
which the path is being refered for more robust data writing.  New test
cases with new changes to the core of how most methods are written to
take arguments.