Completed draft of templates for parallel experiment generation
Taking a similar approach as single experiment driver, but using
function constructors to generate parameter arrays for pmapping to
distributed processing.  Going to test the interface with distributed
as a next step, write up documentation when successfull.  Dropping the
loop slurm submission in favor of robust function wrapper with try
except control flow, and p mapping inside of slurm jobs.